Task: move webDav from AdminBundle to CoreBundle
